requirements.txt
gradio==4.44.0
langchain==0.3.7
langgraph==0.2.40
langchain-openai==0.2.2
langchain-community==0.3.6
pdfminer.six==20231228
httpx==0.27.2
python-dotenv==1.0.1
tiktoken==0.7.0
chromadb==0.5.11
